--- src/main/java/springboot/controller/AuthenticationController.java
@@ -1,25 +1,0 @@
-package springboot.controller;
-
-import jakarta.validation.Valid;
-import lombok.RequiredArgsConstructor;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-import springboot.dto.user.UserRegistrationRequestDto;
-import springboot.dto.user.UserResponseDto;
-import springboot.exeptions.RegistrationException;
-import springboot.service.UserService;
-
-@RestController
-@RequiredArgsConstructor
-@RequestMapping("/auth")
-public class AuthenticationController {
-    private final UserService userService;
-
-    @PostMapping("/registration")
-    public UserResponseDto register(@RequestBody @Valid UserRegistrationRequestDto requestDto)
-            throws RegistrationException {
-        return userService.register(requestDto);
-    }
-}
