+++ src/main/java/springboot/dto/BookDto.java
@@ -1,0 +1,27 @@
+package springboot.dto;
+
+import jakarta.validation.constraints.Min;
+import jakarta.validation.constraints.NotNull;
+import java.math.BigDecimal;
+import lombok.Data;
+
+@Data
+public class BookDto {
+
+    @NotNull
+    private Long id;
+    @NotNull
+    private String title;
+    @NotNull
+    private String author;
+    @NotNull
+    private String isbn;
+    @NotNull
+    @Min(0)
+    private BigDecimal price;
+
+    private String description;
+
+    private String coverImage;
+
+}
