--- src/main/java/springboot/dto/book/CreateBookRequestDto.java
@@ -1,23 +1,0 @@
-package springboot.dto.book;
-
-import jakarta.validation.constraints.Min;
-import jakarta.validation.constraints.NotBlank;
-import jakarta.validation.constraints.NotNull;
-import java.math.BigDecimal;
-import lombok.Data;
-
-@Data
-public class CreateBookRequestDto {
-
-    @NotBlank
-    private String title;
-    @NotBlank
-    private String author;
-    @NotBlank
-    private String isbn;
-    @NotNull
-    @Min(0)
-    private BigDecimal price;
-    private String description;
-    private String coverImage;
-}
