--- src/main/java/springboot/dto/user/UserRegistrationRequestDto.java
@@ -1,21 +1,0 @@
-package springboot.dto.user;
-
-import jakarta.validation.constraints.Email;
-import jakarta.validation.constraints.NotBlank;
-import lombok.Data;
-import org.hibernate.validator.constraints.Length;
-import springboot.validation.FieldMatch;
-
-@FieldMatch(password = "password", repeatedPassword = "confirmPassword")
-@Data
-public class UserRegistrationRequestDto {
-    @NotBlank
-    @Email
-    private String email;
-    @NotBlank
-    @Length(min = 8, max = 20)
-    private String password;
-    @NotBlank
-    @Length(min = 8, max = 20)
-    private String repeatPassword;
-}
