--- src/main/java/springboot/dto/user/UserResponseDto.java
@@ -1,11 +1,0 @@
-package springboot.dto.user;
-
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter
-@Setter
-public class UserResponseDto {
-    private Long id;
-    private String email;
-}
