--- src/main/java/springboot/exeptions/RegistrationException.java
@@ -1,7 +1,0 @@
-package springboot.exeptions;
-
-public class RegistrationException extends RuntimeException {
-    public RegistrationException(String message) {
-        super(message);
-    }
-}
