--- src/main/java/springboot/model/User.java
@@ -1,42 +1,0 @@
-package springboot.model;
-
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import jakarta.persistence.Table;
-import lombok.Getter;
-import lombok.Setter;
-import org.hibernate.annotations.SQLDelete;
-import org.hibernate.annotations.SQLRestriction;
-
-@Getter
-@Setter
-@Table(name = "books")
-@Entity
-@SQLDelete(sql = "UPDATE employees SET is_deleted = true WHERE id=?")
-@SQLRestriction(value = "is_deleted=false")
-public class User {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @Column(nullable = false, unique = true)
-    private String email;
-
-    @Column(nullable = false)
-    private String password;
-
-    @Column(nullable = false)
-    private String firstName;
-
-    @Column(nullable = false)
-    private String lastName;
-
-    private String shippingAddress;
-
-    @Column(nullable = false)
-    private boolean isDeleted = false;
-}
