--- src/main/java/springboot/repository/UserRepository.java
@@ -1,10 +1,0 @@
-package springboot.repository;
-
-import java.util.Optional;
-import org.springframework.data.jpa.repository.JpaRepository;
-import springboot.model.User;
-
-public interface UserRepository extends JpaRepository<User, Long> {
-
-    Optional<User> findByEmail(String email);
-}
