--- src/main/java/springboot/service/UserService.java
@@ -1,8 +1,0 @@
-package springboot.service;
-
-import springboot.dto.user.UserRegistrationRequestDto;
-import springboot.dto.user.UserResponseDto;
-
-public interface UserService {
-    UserResponseDto register(UserRegistrationRequestDto requestDto);
-}
