+++ src/main/java/springboot/service/impl/BookServiceImpl.java
@@ -1,11 +1,10 @@
 package springboot.service.impl;
 
+import java.util.List;
 import lombok.RequiredArgsConstructor;
-import org.springframework.data.domain.Page;
-import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
-import springboot.dto.book.BookDto;
-import springboot.dto.book.CreateBookRequestDto;
+import springboot.dto.BookDto;
+import springboot.dto.CreateBookRequestDto;
 import springboot.exeptions.EntityNotFoundException;
 import springboot.mapper.BookMapper;
 import springboot.model.Book;
@@ -26,9 +25,10 @@ public class BookServiceImpl implements BookService {
     }
 
     @Override
-    public Page<BookDto> getAll(Pageable pageable) {
-        return bookRepository.findAll(pageable)
-                .map(bookMapper::toDto);
+    public List<BookDto> getAll() {
+        return bookRepository.findAll().stream()
+                .map(bookMapper::toDto)
+                .toList();
     }
 
     @Override
