--- src/main/java/springboot/validation/FieldMatch.java
@@ -1,29 +1,0 @@
-package springboot.validation;
-
-import jakarta.validation.Constraint;
-import jakarta.validation.Payload;
-import java.lang.annotation.Documented;
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
-
-
-@Documented
-@Constraint(validatedBy = {FieldMatchValidator.class})
-@Target({ElementType.TYPE, ElementType.PARAMETER})
-@Retention(RetentionPolicy.RUNTIME)
-public @interface FieldMatch {
-    String message() default "Fields is not match";
-    Class<?>[] groups() default {};
-    Class<? extends Payload>[] payload() default {};
-    String password();
-
-    String repeatedPassword();
-
-    @Target({ ElementType.TYPE })
-    @Retention(RetentionPolicy.RUNTIME)
-    @interface List {
-        FieldMatch[] value();
-    }
-}
