--- src/main/java/springboot/validation/FieldMatchValidator.java
@@ -1,26 +1,0 @@
-package springboot.validation;
-
-import jakarta.validation.ConstraintValidator;
-import jakarta.validation.ConstraintValidatorContext;
-import org.springframework.beans.BeanWrapperImpl;
-
-public class FieldMatchValidator implements ConstraintValidator<FieldMatch, String> {
-    private String field;
-    private String fieldMatch;
-
-    public void initialize(FieldMatch constraintAnnotation) {
-        this.field = constraintAnnotation.password();
-        this.fieldMatch = constraintAnnotation.repeatedPassword();
-    }
-
-    @Override
-    public boolean isValid(String value, ConstraintValidatorContext constraintValidatorContext) {
-        Object fieldValue = new BeanWrapperImpl(value)
-                .getPropertyValue(field);
-        Object fieldMatchValue = new BeanWrapperImpl(value)
-                .getPropertyValue(fieldMatch);
-
-        return fieldValue.equals(fieldMatchValue);
-
-    }
-}
